requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
Django==4.2.4
django-debug-toolbar==4.2.0
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
et-xmlfile==1.1.0
inflection==0.5.1
openpyxl==3.1.2
packaging==23.1
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
redis==5.0.0
sqlparse==0.4.4
typing_extensions==4.7.1
uritemplate==4.1.1
